Update naming of the loss stuff
